Follow-up for `java.time` adapter changes
