+++ gson/src/main/java/com/google/gson/internal/bind/TypeAdapters.java
@@ -807,6 +807,8 @@ public final class TypeAdapters {
    * An abstract {@link TypeAdapter} for classes whose JSON serialization consists of a fixed set of
    * integer fields. That is the case for {@link Calendar} and the legacy serialization of various
    * {@code java.time} types.
+   *
+   * <p>This class handles {@code null}; subclasses don't have to use {@link #nullSafe()}.
    */
   abstract static class IntegerFieldsTypeAdapter<T> extends TypeAdapter<T> {
     private final List<String> fields;
@@ -815,8 +817,21 @@ public final class TypeAdapters {
       this.fields = Arrays.asList(fields);
     }
 
+    /**
+     * On deserialization: Creates an object from the integer values. Subclasses should use {@link
+     * Math#toIntExact(long)} and similar if necessary to prevent silent truncation.
+     *
+     * <p>Values have the same order as the field names provided to the {@linkplain
+     * #IntegerFieldsTypeAdapter(String[]) constructor}.
+     */
     abstract T create(long[] values);
 
+    /**
+     * On serialization: Extracts the integer values from the object.
+     *
+     * <p>Values must have the same order as the field names provided to the {@linkplain
+     * #IntegerFieldsTypeAdapter(String[]) constructor}.
+     */
     abstract long[] integerValues(T t);
 
     @Override
@@ -827,7 +842,7 @@ public final class TypeAdapters {
       }
       in.beginObject();
       long[] values = new long[fields.size()];
-      while (in.peek() != JsonToken.END_OBJECT) {
+      while (in.hasNext()) {
         String name = in.nextName();
         int index = fields.indexOf(name);
         if (index >= 0) {
@@ -884,7 +899,7 @@ public final class TypeAdapters {
         }
       };
 
-  // TODO: update this when we are on at least Android API Level 24.
+  // TODO: switch to `Math#toIntExact` when we are on at least Android API Level 24.
   private static int toIntExact(long x) {
     int i = (int) x;
     if (i != x) {
@@ -946,6 +961,10 @@ public final class TypeAdapters {
     TypeAdapterFactory get();
   }
 
+  /**
+   * Adapter factory for {@code java.time} classes. Returns {@code null} if not supported by the
+   * current environment (e.g. too old Android version, without desugaring).
+   */
   public static TypeAdapterFactory javaTimeTypeAdapterFactory() {
     try {
       Class<?> javaTimeTypeAdapterFactoryClass =
